Use last word instead of faulty hash
The last word in each game is guaranteed to be unique. Hence, it is a much
more fail-safe way to notice whether the game in the cache and the one being
loaded differ.

Changing this broke the game, as the values in local storage did not have the
expected values.